refactor(web): extract agent-config + team helpers from web.ts
Move per-agent config getters/setters (model, displayName, securityProfile)
plus the team/hierarchy helpers (TeamConfig, sanitizeTeamConfig,
cleanupTeamReferences) into two new modules:

- src/web/agent-config.ts -- agentDir, readFileOr, model/displayName/
  securityProfile read+write, listAgentNames, isKnownAgent,
  AGENTS_BASE_DIR, MODEL_ALIASES, DEFAULT_MODEL
- src/web/agent-team.ts -- TeamConfig + DEFAULT_TEAM, readAgentTeam,
  writeAgentTeam, sanitizeTeamConfig, cleanupTeamReferences

web.ts: 4956 -> 4729 lines. No behavior change.
